app/requirements.txt
streamlit==1.50.0
tensorflow-cpu==2.11.0
pandas==2.2.2
numpy==1.24.4
scikit-learn==1.5.0
